requirements.txt
pandas==2.0.3
python-dotenv